add utils file and organize some methods
